+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 
 ### Enhancements & Refactors
+- Deprecated the `rate` command [#755](https://github.com/openscope/openscope/issues/755)
+- Add git strategy flow chart to documentation [#80](https://github.com/openscope/openscope/issues/80)